Fix component loading: move to snippets/, add imports, create Formspree version
MAJOR FIXES:
- Move all components from _components/ to snippets/ (Mintlify requirement)
- Add explicit imports to MDX files (required by Mintlify)
- Create Formspree version of ExpertForm (user's preference)
- Add side-by-side comparison with tabs (Formspree vs Supabase)

Changes:
- snippets/ExpertFormFormspree.jsx: New Formspree-based form (ID: xvgeropp)
- snippets/ExpertFormSupabase.jsx: Comparison placeholder
- get-involved/local-expert-connections.mdx: Add imports and tabs
- experts/index.mdx: Add imports
- All components moved from _components/ to snippets/

This fixes the "components not loading" issue - Mintlify requires:
1. Components in snippets/ directory
2. Explicit imports in each MDX file

Sources:
- https://www.mintlify.com/docs/customize/react-components
