+++ experts/index.mdx
@@ -15,6 +15,9 @@ canonical: 'https://travel-info.co.zw/experts'
 'og:type': "website"
 ---
 
+import ExpertDirectory from "/snippets/ExpertDirectory.jsx"
+import TestComponent from "/snippets/TestComponent.jsx"
+
 # Local Expert Directory
 
 Connect with verified local experts who can transform your Zimbabwe experience. Our directory features passionate guides, cultural specialists, and local professionals who have been vetted for knowledge, professionalism, and commitment to authentic, sustainable tourism.
